script: Finish moving all HTML elements to relevant subfolders
This finishes off the remaining elements which didn't have a
clear subfolder. Most of them are actually used as part of
other elements, hence moving them adjacent. Others are textual
and don't have any explicit semantics, other than with CSS.

Fixes #46329

Testing: it compiles
