+++ components/script/dom/globalscope/origin.rs
@@ -19,7 +19,9 @@ use crate::dom::bindings::str::DOMString;
 use crate::dom::globalscope::GlobalScope;
 use crate::dom::html::htmlanchorelement::HTMLAnchorElement;
 use crate::dom::html::htmlareaelement::HTMLAreaElement;
-use crate::dom::html::htmlhyperlinkelementutils::{HyperlinkElement, HyperlinkElementTraits};
+use crate::dom::html::links::htmlhyperlinkelementutils::{
+    HyperlinkElement, HyperlinkElementTraits,
+};
 use crate::dom::url::URL;
 use crate::dom::window::Window;
