+++ components/script/dom/html/document_metadata/htmllinkelement.rs
@@ -57,6 +57,7 @@ use crate::dom::html::document_metadata::processingoptions::{
     LinkFetchContext, LinkFetchContextType, LinkProcessingOptions,
 };
 use crate::dom::html::htmlelement::HTMLElement;
+use crate::dom::html::links::relations::LinkRelations;
 use crate::dom::medialist::MediaList;
 use crate::dom::node::virtualmethods::VirtualMethods;
 use crate::dom::node::{BindContext, Node, NodeTraits, UnbindContext};
@@ -66,7 +67,6 @@ use crate::dom::types::{EventTarget, GlobalScope};
 use crate::fetch::network_listener::{
     FetchResponseListener, ResourceTimingListener, submit_timing,
 };
-use crate::links::LinkRelations;
 use crate::modules::script_module::{ScriptFetchOptions, fetch_a_modulepreload_module};
 use crate::url::ensure_blob_referenced_by_url_is_kept_alive;
