+++ components/script/dom/html/embedded_content/htmlareaelement.rs
@@ -30,10 +30,12 @@ use crate::dom::element::{AttributeMutation, Element, reflect_referrer_policy_at
 use crate::dom::event::Event;
 use crate::dom::eventtarget::EventTarget;
 use crate::dom::html::htmlelement::HTMLElement;
-use crate::dom::html::htmlhyperlinkelementutils::{HyperlinkElement, HyperlinkElementTraits};
+use crate::dom::html::links::htmlhyperlinkelementutils::{
+    HyperlinkElement, HyperlinkElementTraits,
+};
+use crate::dom::html::links::relations::{LinkRelations, follow_hyperlink};
 use crate::dom::node::Node;
 use crate::dom::node::virtualmethods::VirtualMethods;
-use crate::links::{LinkRelations, follow_hyperlink};
 
 #[derive(Debug, PartialEq)]
 pub enum Area {
