+++ components/script/dom/html/grouping_content/mod.rs
@@ -3,10 +3,12 @@
  * file, You can obtain one at https://mozilla.org/MPL/2.0/. */
 
 pub(crate) mod htmldataelement;
+pub(crate) mod htmldirectoryelement;
 pub(crate) mod htmldivelement;
 pub(crate) mod htmldlistelement;
 pub(crate) mod htmlhrelement;
 pub(crate) mod htmllielement;
+pub(crate) mod htmlmenuelement;
 pub(crate) mod htmlolistelement;
 pub(crate) mod htmlparagraphelement;
 pub(crate) mod htmlpreelement;
