+++ components/script/dom/html/mod.rs
@@ -14,6 +14,7 @@ pub(crate) mod grouping_content;
 pub(crate) use self::grouping_content::*;
 pub(crate) mod interactive;
 pub(crate) use self::interactive::*;
+pub(crate) mod links;
 pub(crate) mod scripting;
 pub(crate) use self::scripting::*;
 pub(crate) mod tabular_data;
@@ -21,20 +22,6 @@ pub(crate) use self::tabular_data::*;
 pub(crate) mod textual;
 pub(crate) use self::textual::*;
 pub(crate) mod htmlcollection;
-pub(crate) mod htmldirectoryelement;
 pub(crate) mod htmlelement;
-pub(crate) mod htmlfontelement;
-pub(crate) mod htmlformcontrolscollection;
-pub(crate) mod htmlheadingelement;
-pub(crate) mod htmlhyperlinkelementutils;
-pub(crate) mod htmlmarqueeelement;
-pub(crate) mod htmlmenuelement;
-pub(crate) mod htmlmodelement;
-pub(crate) mod htmloptionscollection;
-pub(crate) mod htmlparamelement;
-pub(crate) mod htmlpictureelement;
-pub(crate) mod htmlquoteelement;
 pub(crate) mod htmlslotelement;
-pub(crate) mod htmlsourceelement;
 pub(crate) mod htmlunknownelement;
-pub(crate) mod interactive_element_command;
